Update of the TPIE library to work with 64-bit data types. This includes:   -  Introduction of TPIE_OS_SIZE_T typedef (0...M-1)      This replaces size_t variables. The exact type     may be set using TPIE_OS_SMALL_MEMORY.   - Consistent use of TPIE_OS_OFFSET typedef (0...N-1)     This replaces off_t, unsinged int, long long etc. used by    various programmers.   - Compile-time-check for potential "loss-of-data" problems    when using 64-bit data types.
git-svn-id: http://svn.daimi.au.dk/svn/tpie/trunk@1201 669acd26-ddd2-4139-9e68-486997aa639d

+++ tpie/include/logstream.h
@@ -4,7 +4,7 @@
 // Author: Darren Erik Vengroff <dev@cs.duke.edu>
 // Created: 5/12/94
 //
-// $Id: logstream.h,v 1.18 2004-05-31 19:31:07 tavi Exp $
+// $Id: logstream.h,v 1.19 2004-08-12 12:35:32 jan Exp $
 //
 
 #ifndef _LOGSTREAM_H
@@ -73,7 +73,7 @@ logstream& operator<< (logstream& o, const logmanip<TP>& m)
   (*m._f)(o, m._a); return o;
 }
 
-logmanip<unsigned int> setpriority(unsigned int p);
-logmanip<unsigned int> setthreshold(unsigned int p);
+logmanip<unsigned long> setpriority(unsigned long p);
+logmanip<unsigned long> setthreshold(unsigned long p);
 
 #endif // _LOGSTREAM_H
